support multi-volume per echo in bids (e.g., functional QSM)
bugfix: Enable input _echo-##_ tags to be read regardless of zero padding use… (#108)

* Enable input _echo-##_ tags to be read regardless of zero padding used in echo number

The original code only matched if the echoes were numbered 1,2,3,4 and would hang in an indefinite loop if these were not found, which is the case for datasets with echoes numbered as 01,02,03,04.

Code now simply matches the echo number, and converts into double, which can then be compared to the actual number used in the loop.

* Fix comment character to matlab %